+++ CHANGELOG.md
@@ -7,12 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-- Faucet
 - Trade
+
+## [0.0.3] - 2024-05-08
+
+### Added
+
+- Allow storing seeds in local file system
+- Coinbase.configure_from_file
+- Faucet
 - Individual private key export
 - Allow disabling debug tracing
 - Error specifications
-
+- WETH ERC-20
 ## [0.0.2] - 2024-05-01
 
 ### Added